requirements.txt
openai>=1.10.0
pydantic>=2.5.0
colorama>=0.4.6
python-dotenv>=1.0.0
# Core dependencies for YEDAN Reactor
praw>=7.7.1
tweepy>=4.14.0
textblob>=0.17.1
schedule>=1.2.0
supabase>=2.3.0
langchain>=0.1.0
playwright>=1.40.0
browser-use>=0.1.0
# V1300 Toolchain Supremacy
crawl4ai>=0.5.0
pydantic-ai>=0.1.0
httpx>=0.28.0

